cast

+++ a3/q2main.cc
@@ -17,7 +17,7 @@ void inputpoly(char *fileA, char *fileB){
             int *poly1 = new int[size1];
             int *poly2 = new int[size2];
             int *polyr = new int[sizer]{};
-            for(int i=0; i<size1; ++i){//assume sizes are the same
+            for(unsigned int i=0; i<size1; ++i){//assume sizes are the same
                 file1 >> poly1[i];
                 file2 >> poly2[i];
             }
@@ -75,7 +75,7 @@ void randompoly(long numCoef){
         poly2[i] = i+1;
     }
     Poly c = {polyr, sizer};
-    polymultiply(Poly{poly1, numCoef}, Poly{poly2, numCoef}, c, floor(sqrt(sizer)));
+    polymultiply(Poly{poly1, (size_t)numCoef}, Poly{poly2, (size_t)numCoef}, c, floor(sqrt(sizer)));
 }
 
 int main( int argc, char * argv[] ) {
